+++ adminalerts/tests/test_permissions_ajax.py
@@ -9,8 +9,10 @@ from projectroles.tests.test_permissions import TestSiteAppPermissionBase
 from adminalerts.tests.test_models import AdminAlertMixin
 
 
-class TestAdminAlertPermissions(AdminAlertMixin, TestSiteAppPermissionBase):
-    """Tests for AdminAlert views"""
+class TestAdminAlertActiveToggleAjaxView(
+    AdminAlertMixin, TestSiteAppPermissionBase
+):
+    """Permission tests for AdminAlertActiveToggleAjaxView"""
 
     def setUp(self):
         super().setUp()
@@ -21,23 +23,19 @@ class TestAdminAlertPermissions(AdminAlertMixin, TestSiteAppPermissionBase):
             description='description',
             active=True,
         )
-
-    def test_active_toggle(self):
-        """Test permissions for activation Ajax view"""
-        url = reverse(
+        self.url = reverse(
             'adminalerts:ajax_active_toggle',
             kwargs={'adminalert': self.alert.sodar_uuid},
         )
+
+    def test_post(self):
+        """Test AdminAlertActiveToggleAjaxView POST"""
         good_users = [self.superuser]
         bad_users = [self.anonymous, self.regular_user]
-        self.assert_response(url, good_users, 200, method='POST')
-        self.assert_response(url, bad_users, 403, method='POST')
+        self.assert_response(self.url, good_users, 200, method='POST')
+        self.assert_response(self.url, bad_users, 403, method='POST')
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_active_toggle_anon(self):
-        """Test permissions for activation Ajax view with anonymous access"""
-        url = reverse(
-            'adminalerts:ajax_active_toggle',
-            kwargs={'adminalert': self.alert.sodar_uuid},
-        )
-        self.assert_response(url, self.anonymous, 403, method='POST')
+    def test_post_anon(self):
+        """Test POST with anonymous access"""
+        self.assert_response(self.url, self.anonymous, 403, method='POST')