+++ timeline/tests/test_permissions.py
@@ -1,4 +1,4 @@
-"""UI view permission tests for the timeline app"""
+"""Tests for UI view permissions in the timeline app"""
 
 from django.test import override_settings
 from django.urls import reverse
@@ -7,14 +7,17 @@ from django.urls import reverse
 from projectroles.tests.test_permissions import TestProjectPermissionBase
 
 
-class TestTimelinePermissions(TestProjectPermissionBase):
-    """Tests for timeline views"""
+class TestProjectTimelineView(TestProjectPermissionBase):
+    """Tests for ProjectTimelineView permissions"""
 
-    def test_project_list(self):
-        """Test project event list"""
-        url = reverse(
+    def setUp(self):
+        super().setUp()
+        self.url = reverse(
             'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
         )
+
+    def test_get(self):
+        """Test ProjectTimelineView GET"""
         good_users = [
             self.superuser,
             self.user_owner_cat,
@@ -27,25 +30,49 @@ class TestTimelinePermissions(TestProjectPermissionBase):
             self.user_guest,
         ]
         bad_users = [self.user_finder_cat, self.user_no_roles, self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-        # Test public project
+        self.assert_response(self.url, good_users, 200)
+        self.assert_response(self.url, bad_users, 302)
         self.project.set_public()
-        self.assert_response(url, self.user_no_roles, 200)
-        self.assert_response(url, self.anonymous, 302)
+        self.assert_response(self.url, self.user_no_roles, 200)
+        self.assert_response(self.url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_project_list_anon(self):
-        """Test project event list with anonynomus access"""
-        url = reverse(
-            'timeline:list_project', kwargs={'project': self.project.sodar_uuid}
-        )
+    def test_get_anon(self):
+        """Test GET with anonynomus access"""
         self.project.set_public()
-        self.assert_response(url, self.anonymous, 200)
+        self.assert_response(self.url, self.anonymous, 200)
 
-    def test_site_list(self):
-        """Test site event list"""
-        url = reverse('timeline:list_site')
+    def test_get_archive(self):
+        """Test GET with archived project"""
+        self.project.set_archive()
+        good_users = [
+            self.superuser,
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,
+            self.user_guest,
+        ]
+        bad_users = [self.user_finder_cat, self.user_no_roles, self.anonymous]
+        self.assert_response(self.url, good_users, 200)
+        self.assert_response(self.url, bad_users, 302)
+        self.project.set_public()
+        self.assert_response(self.url, self.user_no_roles, 200)
+        self.assert_response(self.url, self.anonymous, 302)
+
+
+class TestSiteTimelineView(TestProjectPermissionBase):
+    """Tests for SiteTimelineView permissions"""
+
+    def setUp(self):
+        super().setUp()
+        self.url = reverse('timeline:list_site')
+
+    def test_get(self):
+        """Test SiteTimelineView GET"""
         good_users = [
             self.superuser,
             self.user_owner_cat,
@@ -59,21 +86,25 @@ class TestTimelinePermissions(TestProjectPermissionBase):
             self.user_guest,
             self.user_no_roles,
         ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, self.anonymous, 302)
+        self.assert_response(self.url, good_users, 200)
+        self.assert_response(self.url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_stite_list_anon(self):
-        """Test site event list with anonynomus access"""
-        url = reverse('timeline:list_site')
-        self.assert_response(url, self.anonymous, 302)
+    def test_get_anon(self):
+        """Test GET with anonynomus access"""
+        self.assert_response(self.url, self.anonymous, 302)
 
-    def test_admin_list(self):
-        """Test admin event list"""
-        url = reverse('timeline:timeline_site_admin')
-        good_users = [
-            self.superuser,
-        ]
+
+class TestAdminTimelineView(TestProjectPermissionBase):
+    """Tests for AdminTimelineView permissions"""
+
+    def setUp(self):
+        super().setUp()
+        self.url = reverse('timeline:timeline_site_admin')
+
+    def test_get(self):
+        """Test AdminTimelineView GET"""
+        good_users = [self.superuser]
         bad_users = [
             self.user_owner_cat,
             self.user_delegate_cat,
@@ -87,18 +118,21 @@ class TestTimelinePermissions(TestProjectPermissionBase):
             self.user_no_roles,
             self.anonymous,
         ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(self.url, good_users, 200)
+        self.assert_response(self.url, bad_users, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_admin_list_anon(self):
-        """Test admin list with anonymous access"""
-        url = reverse('timeline:timeline_site_admin')
-        self.assert_response(url, self.anonymous, 302)
+    def test_get_anon(self):
+        """Test GET with anonymous access"""
+        self.assert_response(self.url, self.anonymous, 302)
 
-    def test_project_event_object_list(self):
-        """Test project event object list"""
-        url = reverse(
+
+class TestProjectObjectTimelineView(TestProjectPermissionBase):
+    """Tests for ProjectObjectTimelineView permissions"""
+
+    def setUp(self):
+        super().setUp()
+        self.url = reverse(
             'timeline:list_object',
             kwargs={
                 'project': self.project.sodar_uuid,
@@ -106,6 +140,9 @@ class TestTimelinePermissions(TestProjectPermissionBase):
                 'object_uuid': self.user_owner.sodar_uuid,
             },
         )
+
+    def test_get(self):
+        """Test ProjectObjectTimelineView GET"""
         good_users = [
             self.superuser,
             self.user_owner_cat,
@@ -118,36 +155,55 @@ class TestTimelinePermissions(TestProjectPermissionBase):
             self.user_guest,
         ]
         bad_users = [self.user_finder_cat, self.user_no_roles, self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-        # Test public project
+        self.assert_response(self.url, good_users, 200)
+        self.assert_response(self.url, bad_users, 302)
         self.project.set_public()
-        self.assert_response(url, self.user_no_roles, 200)
-        self.assert_response(url, self.anonymous, 302)
+        self.assert_response(self.url, self.user_no_roles, 200)
+        self.assert_response(self.url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
     def test_project_event_object_list_anon(self):
-        """Test object event list with anonynomus access"""
-        url = reverse(
-            'timeline:list_object',
-            kwargs={
-                'project': self.project.sodar_uuid,
-                'object_model': 'User',
-                'object_uuid': self.user_owner.sodar_uuid,
-            },
-        )
+        """Test GET with anonynomus access"""
         self.project.set_public()
-        self.assert_response(url, self.anonymous, 200)
+        self.assert_response(self.url, self.anonymous, 200)
 
-    def test_site_event_object_list(self):
-        """Test site event object list"""
-        url = reverse(
+    def test_get_archive(self):
+        """Test GET with archived project"""
+        self.project.set_archive()
+        good_users = [
+            self.superuser,
+            self.user_owner_cat,
+            self.user_delegate_cat,
+            self.user_contributor_cat,
+            self.user_guest_cat,
+            self.user_owner,
+            self.user_delegate,
+            self.user_contributor,
+            self.user_guest,
+        ]
+        bad_users = [self.user_finder_cat, self.user_no_roles, self.anonymous]
+        self.assert_response(self.url, good_users, 200)
+        self.assert_response(self.url, bad_users, 302)
+        self.project.set_public()
+        self.assert_response(self.url, self.user_no_roles, 200)
+        self.assert_response(self.url, self.anonymous, 302)
+
+
+class TestSiteObjectTimelineView(TestProjectPermissionBase):
+    """Tests for SiteObjectTimelineView permissions"""
+
+    def setUp(self):
+        super().setUp()
+        self.url = reverse(
             'timeline:list_object_site',
             kwargs={
                 'object_model': 'User',
                 'object_uuid': self.user_owner.sodar_uuid,
             },
         )
+
+    def test_get(self):
+        """Test SiteObjectTimelineView GET"""
         good_users = [
             self.superuser,
             self.user_owner_cat,
@@ -161,22 +217,14 @@ class TestTimelinePermissions(TestProjectPermissionBase):
             self.user_guest,
             self.user_no_roles,
         ]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, self.anonymous, 302)
-        # Test public project
+        self.assert_response(self.url, good_users, 200)
+        self.assert_response(self.url, self.anonymous, 302)
         self.project.set_public()
-        self.assert_response(url, self.user_no_roles, 200)
-        self.assert_response(url, self.anonymous, 302)
+        self.assert_response(self.url, self.user_no_roles, 200)
+        self.assert_response(self.url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_site_event_object_list_anon(self):
-        """Test site event list with anonymous access"""
-        url = reverse(
-            'timeline:list_object_site',
-            kwargs={
-                'object_model': 'User',
-                'object_uuid': self.user_owner.sodar_uuid,
-            },
-        )
+    def test_get_anon(self):
+        """Test GET with anonymous access"""
         self.project.set_public()
-        self.assert_response(url, self.anonymous, 302)
+        self.assert_response(self.url, self.anonymous, 302)