+++ tokens/tests/test_permissions.py
@@ -1,4 +1,4 @@
-"""Permission tests for the tokens app"""
+"""Tests for UI view permissions in the tokens app"""
 
 from django.test import override_settings
 from django.urls import reverse
@@ -12,45 +12,35 @@ from projectroles.tests.test_permissions import TestSiteAppPermissionBase
 class TestTokenPermissions(TestSiteAppPermissionBase):
     """Tests for token view permissions"""
 
-    def test_list(self):
-        """Test permissions for token list"""
+    def test_get_list(self):
+        """Test tUserTokenListView GET"""
         url = reverse('tokens:list')
-        good_users = [self.superuser, self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, [self.superuser, self.regular_user], 200)
+        self.assert_response(url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_list_anon(self):
-        """Test permissions for token list with anonymous access"""
+    def test_get_list_anon(self):
+        """Test UserTokenListView GET with anonymous access"""
         url = reverse('tokens:list')
-        good_users = [self.superuser, self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, [self.superuser, self.regular_user], 200)
+        self.assert_response(url, self.anonymous, 302)
 
-    def test_create(self):
-        """Test permissions for token creation"""
+    def test_get_create(self):
+        """Test UserTokenCreateView GET"""
         url = reverse('tokens:create')
-        good_users = [self.superuser, self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, [self.superuser, self.regular_user], 200)
+        self.assert_response(url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_create_anon(self):
-        """Test permissions for token creation with anonymous access"""
+    def test_get_create_anon(self):
+        """Test UserTokenCreateView GET with anonymous access"""
         url = reverse('tokens:create')
-        good_users = [self.superuser, self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, [self.superuser, self.regular_user], 200)
+        self.assert_response(url, self.anonymous, 302)
 
-    def test_delete(self):
-        """Test permissions for token deletion"""
+    def test_get_delete(self):
+        """Test UserTokenDeleteView GET"""
         token = AuthToken.objects.create(self.regular_user, None)
         url = reverse('tokens:delete', kwargs={'pk': token[0].pk})
-        good_users = [self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, self.regular_user, 200)
+        self.assert_response(url, self.anonymous, 302)