+++ userprofile/tests/test_permissions.py
@@ -1,4 +1,4 @@
-"""Tests for permissions in the userprofile app"""
+"""Tests for UI view permissions in the userprofile app"""
 
 from django.test import override_settings
 from django.urls import reverse
@@ -10,36 +10,28 @@ from projectroles.tests.test_permissions import TestSiteAppPermissionBase
 class TestUserProfilePermissions(TestSiteAppPermissionBase):
     """Tests for userprofile view permissions"""
 
-    def test_profile(self):
-        """Test permissions for user profile view"""
+    def test_get_profile(self):
+        """Test UserDetailView GET"""
         url = reverse('userprofile:detail')
-        good_users = [self.superuser, self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, [self.superuser, self.regular_user], 200)
+        self.assert_response(url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_profile_anon(self):
-        """Test permissions for user profile view with anonymous access"""
+    def test_get_profile_anon(self):
+        """Test UserDetailView GET with anonymous access"""
         url = reverse('userprofile:detail')
-        good_users = [self.superuser, self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, [self.superuser, self.regular_user], 200)
+        self.assert_response(url, self.anonymous, 302)
 
-    def test_settings_update(self):
-        """Test permissions for user settings update view"""
+    def test_get_settings_update(self):
+        """Test UserSettingUpdateView GET"""
         url = reverse('userprofile:settings_update')
-        good_users = [self.superuser, self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, [self.superuser, self.regular_user], 200)
+        self.assert_response(url, self.anonymous, 302)
 
     @override_settings(PROJECTROLES_ALLOW_ANONYMOUS=True)
-    def test_settings_update_anon(self):
-        """Test permissions for update view with anonymous access"""
+    def test_get_settings_update_anon(self):
+        """Test UserSettingUpdateView GET with anonymous access"""
         url = reverse('userprofile:settings_update')
-        good_users = [self.superuser, self.regular_user]
-        bad_users = [self.anonymous]
-        self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
+        self.assert_response(url, [self.superuser, self.regular_user], 200)
+        self.assert_response(url, self.anonymous, 302)